moved multisim exceptions
