+++ coraplex/src/coraplex/orm/ormatic_interface.py
@@ -7968,7 +7968,7 @@ class MujocoBuilderDAO(
 
 
 class MultiSimErrorDAO(
-    Base, DataAccessObject[semantic_digital_twin.adapters.multi_sim.MultiSimError]
+    Base, DataAccessObject[semantic_digital_twin.exceptions.MultiSimError]
 ):
     __tablename__ = "MultiSimErrorDAO"
 
@@ -7987,8 +7987,7 @@ class MultiSimErrorDAO(
 
 
 class MujocoErrorDAO(
-    MultiSimErrorDAO,
-    DataAccessObject[semantic_digital_twin.adapters.multi_sim.MujocoError],
+    MultiSimErrorDAO, DataAccessObject[semantic_digital_twin.exceptions.MujocoError]
 ):
     __tablename__ = "MujocoErrorDAO"
 
@@ -8006,9 +8005,7 @@ class MujocoErrorDAO(
 
 class MujocoEntityNotFoundErrorDAO(
     MujocoErrorDAO,
-    DataAccessObject[
-        semantic_digital_twin.adapters.multi_sim.MujocoEntityNotFoundError
-    ],
+    DataAccessObject[semantic_digital_twin.exceptions.MujocoEntityNotFoundError],
 ):
     __tablename__ = "MujocoEntityNotFoundErrorDAO"
 
@@ -11420,9 +11417,7 @@ class QoSProfileJSONSerializerDAO(
 
 class QuaternionConversionErrorDAO(
     MultiSimErrorDAO,
-    DataAccessObject[
-        semantic_digital_twin.adapters.multi_sim.QuaternionConversionError
-    ],
+    DataAccessObject[semantic_digital_twin.exceptions.QuaternionConversionError],
 ):
     __tablename__ = "QuaternionConversionErrorDAO"
