+++ experiments/src/experiments/orm/ormatic_interface.py
@@ -8103,7 +8103,7 @@ class MujocoBuilderDAO(
 
 
 class MultiSimErrorDAO(
-    Base, DataAccessObject[semantic_digital_twin.adapters.multi_sim.MultiSimError]
+    Base, DataAccessObject[semantic_digital_twin.exceptions.MultiSimError]
 ):
     __tablename__ = "MultiSimErrorDAO"
 
@@ -8122,8 +8122,7 @@ class MultiSimErrorDAO(
 
 
 class MujocoErrorDAO(
-    MultiSimErrorDAO,
-    DataAccessObject[semantic_digital_twin.adapters.multi_sim.MujocoError],
+    MultiSimErrorDAO, DataAccessObject[semantic_digital_twin.exceptions.MujocoError]
 ):
     __tablename__ = "MujocoErrorDAO"
 
@@ -8141,9 +8140,7 @@ class MujocoErrorDAO(
 
 class MujocoEntityNotFoundErrorDAO(
     MujocoErrorDAO,
-    DataAccessObject[
-        semantic_digital_twin.adapters.multi_sim.MujocoEntityNotFoundError
-    ],
+    DataAccessObject[semantic_digital_twin.exceptions.MujocoEntityNotFoundError],
 ):
     __tablename__ = "MujocoEntityNotFoundErrorDAO"
 
@@ -11840,9 +11837,7 @@ class QoSProfileJSONSerializerDAO(
 
 class QuaternionConversionErrorDAO(
     MultiSimErrorDAO,
-    DataAccessObject[
-        semantic_digital_twin.adapters.multi_sim.QuaternionConversionError
-    ],
+    DataAccessObject[semantic_digital_twin.exceptions.QuaternionConversionError],
 ):
     __tablename__ = "QuaternionConversionErrorDAO"
